Add mouse click actions
Make some of the arcade buttons dual purpose by generating mouse click
events in addition to button presses. This may be a little hacky, but it
is a reasonable first step. However, this commit exposes a problem with
the way event reports are handled. Frequently reports get lost which
results in lost button presses. This will need to be fixed, probably by
explicitly waiting for the previous USB report to complete sending
before sending the next one. The code as it stands right now is flat out
racy.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>