Enable all 4 NeoPixel channels
Enable driving NeoPixels on all 4 NeoPixel outputs, aka all 4 output
channels of TIM1 in PWM mode. The TIM device supports writing to
multiple registers on each DMA trigger. The driver uses this to load the
transmit buffer with PWM values for each of the output pins.

There is still only one buffer for NeoPixel colour data, but now it is
divided into 4, with each channel assigned a quarter of the buffer. So,
Channel 1 gets the first 1/4, Channel 2 gets the second 1/4, etc.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>