Add a simple RGB animations engine
Add a simple animation engine for the Neopixels. The engine only knows
how to do two things. It can display each of the player's assigned
colours, or it can cycle through the rainbow in a variety of patterns.
Changing modes is accomplished by pressing the P4 "start & select"
buttons at the same time.

There is lots of room to improve the display effects, and it would look
better if all the effects could take into account the assigned player
colours and the trackball position. It also needs to be extended to
accept output reports from the host so that the game software can set
the colours to what it wants.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>